稍微统一了一点ecs_test的shader，但是shadow shader依然是430 core

+++ tests/ecs_test/glsl/blur_vs.glsl
@@ -1,5 +1,5 @@
-#version 430 core
-precision highp float;
+#version 320 es
+precision mediump float;
 
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec2 aTexCoords;
